requirements.txt
openai==0.27.2
python-dotenv==1.0.0
google_search_results==2.4.2
colorama==0.4.6
bs4
selenium
tiktoken==0.3.3
pinecone-client==2.2.1
